update to 1.1.0 remove unused chapter setting fix title match mode adjust ui for easy usage alert on merge errors

+++ src/main/kotlin/com/presisco/mkvbatchop/ui/TrackConfigBox.kt
@@ -143,7 +143,7 @@ class TrackConfigBox(
         conf["offset"]?.let {
             val offset = it as Number
             offsetCheckBox.isSelected = true
-            offsetText.text = offset.toString()
+            offsetText.text = offset.toInt().toString()
 
         }?: clearOffset()
     }
